+++ Net_Core/BusinessLogic/BDirectories.cs
@@ -4,6 +4,7 @@ using GrowthWare.Framework.Models;
 using System;
 using System.Collections.ObjectModel;
 using System.Data;
+using System.Threading.Tasks;
 
 namespace GrowthWare.BusinessLogic;
 
@@ -86,12 +87,12 @@ public class BDirectories : AbstractBusinessLogic
     /// Gets the directories.
     /// </summary>
     /// <returns>Collection{MDirectoryProfile}.</returns>
-    public Collection<MDirectoryProfile> Directories()
+    public async Task<Collection<MDirectoryProfile>> Directories()
     {
         Collection<MDirectoryProfile> mRetVal = new Collection<MDirectoryProfile>();
         if (DatabaseIsOnline()) 
         {
-            DataTable mDataTable = m_DDirectories.Directories();
+            DataTable mDataTable = await m_DDirectories.Directories();
             foreach (DataRow mDataRow in mDataTable.Rows)
             {
                 MDirectoryProfile mProfile = new MDirectoryProfile(mDataRow);
@@ -105,8 +106,8 @@ public class BDirectories : AbstractBusinessLogic
     /// Saves the specified profile.
     /// </summary>
     /// <param name="profile">The profile.</param>
-    public void Save(MDirectoryProfile profile)
+    public async Task Save(MDirectoryProfile profile)
     {
-        if(DatabaseIsOnline()) m_DDirectories.Save(profile);
+        if(DatabaseIsOnline()) await m_DDirectories.Save(profile);
     }
 }