+++ Net_Core/Web.Support/BaseControllers/AbstractFunctionController.cs
@@ -69,7 +69,7 @@ public abstract class AbstractFunctionController : ControllerBase
 
     [AllowAnonymous]
     [HttpGet("GetFunction")]
-    public ActionResult<UIFunctionProfile> GetFunctionForEdit(int functionSeqId)
+    public async Task<ActionResult<UIFunctionProfile>> GetFunctionForEdit(int functionSeqId)
     {
         MSecurityInfo mSecurityInfo = this.getSecurityInfo("FunctionSecurity");
         if(mSecurityInfo != null && mSecurityInfo.MayView)
@@ -84,7 +84,7 @@ public abstract class AbstractFunctionController : ControllerBase
             UIFunctionProfile mRetVal = new UIFunctionProfile(mFunctionProfile);
             mRetVal.CanSaveGroups = this.getSecurityInfo("View_Function_Group_Tab").MayView;
             mRetVal.CanSaveRoles = this.getSecurityInfo("View_Function_Role_Tab").MayView;
-            mRetVal.DirectoryData = DirectoryUtility.GetDirectoryProfile(mFunctionProfile.Id);
+            mRetVal.DirectoryData = await DirectoryUtility.GetDirectoryProfile(mFunctionProfile.Id);
             mRetVal.DirectoryData.ImpersonatePassword = string.Empty; // We don't want the password to show
             mRetVal.FunctionMenuOrders = FunctionUtility.GetFunctionOrder(mFunctionProfile.Id);
             return Ok(mRetVal);
@@ -153,7 +153,7 @@ public abstract class AbstractFunctionController : ControllerBase
 
     [AllowAnonymous]
     [HttpPost("Save")]
-    public ActionResult<bool> Save(UIFunctionProfile functionProfile)
+    public async Task<ActionResult<bool>> Save(UIFunctionProfile functionProfile)
     {
         MSecurityInfo mSecurityInfo = this.getSecurityInfo("FunctionSecurity");
         MSecurityInfo mViewRoleTabSecurityInfo = this.getSecurityInfo("View_Function_Role_Tab");
@@ -248,7 +248,7 @@ public abstract class AbstractFunctionController : ControllerBase
                 mProfileToSave.Id = mFunctionSeqId;
                 if(!string.IsNullOrWhiteSpace(functionProfile.DirectoryData.Directory))
                 {
-                    MDirectoryProfile mDirectoryProfile = DirectoryUtility.GetDirectoryProfile(mProfileToSave.Id);
+                    MDirectoryProfile mDirectoryProfile = await DirectoryUtility.GetDirectoryProfile(mProfileToSave.Id);
                     if(mDirectoryProfile == null)
                     {
                         mDirectoryProfile = new MDirectoryProfile();
@@ -266,7 +266,7 @@ public abstract class AbstractFunctionController : ControllerBase
                     }
                     mDirectoryProfile.Directory = functionProfile.DirectoryData.Directory;
                     mDirectoryProfile.UpdatedBy = mRequestingProfile.Id;
-                    DirectoryUtility.Save(mDirectoryProfile);
+                    await DirectoryUtility.Save(mDirectoryProfile);
                 }
                 string mCommaSeporatedIds = String.Join(",", functionProfile.FunctionMenuOrders.Select(item => item.Function_Seq_Id.ToString()));
                 FunctionUtility.UpdateMenuOrder(mCommaSeporatedIds, mProfileToSave);