+++ Net_Core/Web.Support/Utilities/DirectoryUtility.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.ObjectModel;
 using System.Globalization;
 using System.Linq;
+using System.Threading.Tasks;
 using GrowthWare.BusinessLogic;
 using GrowthWare.Framework;
 using GrowthWare.Framework.Models;
@@ -25,7 +26,7 @@ public static class DirectoryUtility
     /// Gets the directories.
     /// </summary>
     /// <returns>Collection{MDirectoryProfile}.</returns>
-    public static Collection<MDirectoryProfile> Directories()
+    public static async Task<Collection<MDirectoryProfile>> Directories()
     {
         MSecurityEntity mSecurityEntityProfile = SecurityEntityUtility.CurrentProfile;
         String mCacheName = mSecurityEntityProfile.Id.ToString(CultureInfo.InvariantCulture) + "_" + s_DirectoryInfoCachedName;
@@ -33,7 +34,7 @@ public static class DirectoryUtility
         if(mRetVal == null)
         {
             BDirectories mBDirectories = BusinessLogic;
-            mRetVal = mBDirectories.Directories();
+            mRetVal = await mBDirectories.Directories();
             m_CacheHelper.AddToCache(mCacheName, mRetVal);
         }
         return mRetVal;
@@ -60,31 +61,32 @@ public static class DirectoryUtility
     /// </summary>
     /// <param name="functionSeqId">int</param>
     /// <returns>A populated MDirectoryProfile or null</returns>
-    public static MDirectoryProfile GetDirectoryProfile(int functionSeqId)
+    public static async Task<MDirectoryProfile> GetDirectoryProfile(int functionSeqId)
     {
-            var mResult = from mProfile in Directories()
-                          where mProfile.Id == functionSeqId
-                          select mProfile;
-            MDirectoryProfile mRetVal = null;
-            try
-            {
-                mRetVal = mResult.First();
-            }
-            catch (NullReferenceException)
-            {
-                mRetVal = new MDirectoryProfile();
-            }
-            catch (InvalidOperationException)
-            {
-                return new MDirectoryProfile();
-            }
-            return mRetVal;
+        Collection<MDirectoryProfile> mDirectories = await Directories();
+        var mResult = from mProfile in mDirectories
+                        where mProfile.Id == functionSeqId
+                        select mProfile;
+        MDirectoryProfile mRetVal = null;
+        try
+        {
+            mRetVal = mResult.First();
+        }
+        catch (NullReferenceException)
+        {
+            mRetVal = new MDirectoryProfile();
+        }
+        catch (InvalidOperationException)
+        {
+            return new MDirectoryProfile();
+        }
+        return mRetVal;
     }
 
-    public static void Save(MDirectoryProfile profile)
+    public static async Task Save(MDirectoryProfile profile)
     {
         BDirectories mBDirectories = BusinessLogic;
-        mBDirectories.Save(profile);
+        await mBDirectories.Save(profile);
         String mCacheName = SecurityEntityUtility.CurrentProfile.Id.ToString(CultureInfo.InvariantCulture) + "_" + s_DirectoryInfoCachedName;
         m_CacheHelper.RemoveFromCache(mCacheName);
     }