+++ Release_Notes.md
@@ -16,6 +16,7 @@
     - DLogging (ILogging, BLogger, LoggingUtility, AbstractLoggingController)
     - DMessages (IMessages, BMessages, MessagesUtility, AbstractMessagesController)
     - DNameValuePairs (INameValuePairs, BNameValuePairs, NameValuePairsUtility, AbstractNameValuePairsController)
+    - DDirectories (IDirectories, BDirectories, DirectoriesUtility, AbstractDirectoriesController)
 
 ### Known Bugs
 - There is an issue where the refresh token is being revoked when it shouldn't be causing a logout to occure